requirements.txt
openai
langchain_text_splitters
anthropic
chromadb
boto3
numpy
aiohttp
asyncio
fastapi
uvicorn
pydantic
aiofiles
tiktoken
pillow
requests
python-dotenv
gitpython